utils: Add utility module for simplifying common patterns
The first such utility is a macro that expands into a g-expression for
calling into a program from a given guix package. This g-expression
specifically will *not* execl the binary so that only nonzero results
will cause the hook to exit. This is useful for composing with other
hooks (e.g. running several programs in sequence).

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>