library: Add gitlint/commit-msg
This runs gitlint as a commit-msg hooks to lint your git commit
message. If the commit message violates the gitlint rules, the user
has the option of re-editing it or accepting as-is.

The 'program' macro cannot be used directly since we need to point
stdin at /dev/tty in order to accept keyboard input to prompt the user
on lint failure.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>