core: Build g-hooks as a guix extension
Instead of having a wrapper script that uses guix as a library,
implement g-hooks as a guix extension. Using a guix extensions has a
few major advantages:

- The user's custom channels will be used by default

- The user's current guix generation will be used instead of whatever
  is packaged by guix

- Command-line parsing can be handled by guix

The pre-inst-env.sh script is modified so that it will add the correct
path to GUIX_EXTENSIONS_PATH. Note that the program installation will
need to change; I plan on adding a Makefile to handle this soon.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>