library: Add gerrit commit-msg hook
Gerrit's commit-msg hook adds a Change-Id: trailer to the end of your
commit message so it can be used with gerrit's code review system. Add
this hook to g-hooks using the bash-program* wrapper.

The hook itself is vendored directly from gerrit sources; this will
have to be updated as changes are made. The only change I've made is
modifying the license header to follow the REUSE specification.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>